refactor: moved types and constants in different file
