+++ src/constants.ts
@@ -1,0 +1,12 @@
+export const windowIdleEvents = ['scroll', 'resize']
+export const documentIdleEvents = [
+  'wheel',
+  'keydown',
+  'keyup',
+  'mousedown',
+  'mousemove',
+  'touchstart',
+  'touchmove',
+  'click',
+  'contextmenu',
+]
